backend/chat/requirements.txt
djangorestframework~=3.11.1
django-filter~=2.4.0
drf-yasg
django-cors-headers
pymongo==3.11.0
djongo==1.3.3
sqlparse==0.2.4
djoser
common-structure-microservices
django-background-tasks
Django~=3.0.5
dnspython
exrex~=0.10.5
munch~=2.5.0
rstr
channels
channels-redis
redis
py_eureka_client==0.7.4
requests